move modules examples into `examples/modules` folder
