--- src/modules/index.ts
@@ -1,2 +1,0 @@
-export * from './compiler.module'
-export * from './resolver.module'
